refactor(test-framework): convert TestSuite to header-only
Move all implementations from test_framework.cpp into the
class body (implicitly inline). Delete the .cpp file.
